+++ moecog/datasets/__init__.py
@@ -1,6 +1,7 @@
 """ECoG dataset loaders."""
 
 from .base import BaseECoGDataset, ElectrodeInfo
+from .bids import BIDSiEEGDataset, FilmIEEG, HermesVisualECoG, PodcastECoG, VisualECoG
 from .fake import FakeECoGDataset
 from .miller_library import EXPERIMENTS, MillerLibrary, annotations_from_codes
 
@@ -11,4 +12,9 @@ __all__ = [
     "MillerLibrary",
     "EXPERIMENTS",
     "annotations_from_codes",
+    "BIDSiEEGDataset",
+    "HermesVisualECoG",
+    "PodcastECoG",
+    "FilmIEEG",
+    "VisualECoG",
 ]